requirements.txt
tweepy==4.14.0
python-dateutil==2.9.0.post0
pytz==2024.1
